Conway's rules to cell life
paints the cells to the frame if they are alive or dead.  This is determined by conway's series of rules.
if the cell has neighbors that are greater than 4 or less than 2.  It dies
if it has two neighbors then it is unaffected.
if it has three neighbors it comes to life or stays alive.